backend/requirements.txt
fastapi
uvicorn[standard]
httpx
web3
eth-account
python-dotenv
reportlab
PyPDF2
python-docx
qrcode[pil]
alchemy-sdk
supabase
